Site Editor Navigating test: fix switching to grid and waiting for editor
