+++ packages/e2e-test-utils-playwright/src/admin/index.ts
@@ -11,6 +11,7 @@ import { getPageError } from './get-page-error';
 import { visitAdminPage } from './visit-admin-page';
 import { editPost } from './edit-post';
 import { visitSiteEditor } from './visit-site-editor';
+import { waitForSiteEditor } from './wait-for-site-editor';
 import type { PageUtils } from '../page-utils';
 import type { Editor } from '../editor';
 
@@ -45,4 +46,7 @@ export class Admin {
 	visitAdminPage: typeof visitAdminPage = visitAdminPage.bind( this );
 	/** @borrows visitSiteEditor as this.visitSiteEditor */
 	visitSiteEditor: typeof visitSiteEditor = visitSiteEditor.bind( this );
+	/** @borrows waitForSiteEditor as this.waitForSiteEditor */
+	waitForSiteEditor: typeof waitForSiteEditor =
+		waitForSiteEditor.bind( this );
 }
