+++ packages/e2e-test-utils-playwright/src/admin/visit-site-editor.ts
@@ -57,44 +57,6 @@ export async function visitSiteEditor(
 	 * `waitWhileSiteEditorLoading` function in the `edit-site` package.
 	 */
 	if ( ! query.size || postId || canvas === 'edit' ) {
-		await this.page.evaluate( () => {
-			const MAX_LOADING_TIME = 10000;
-			const MAX_PAUSE_TIME = 100;
-
-			return new Promise< void >( ( resolve ) => {
-				let pauseTimeout: ReturnType< typeof setTimeout > | undefined;
-
-				function finish() {
-					unsubscribe();
-					clearTimeout( pauseTimeout );
-					clearTimeout( maxTimeout );
-					resolve();
-				}
-
-				const maxTimeout = setTimeout( finish, MAX_LOADING_TIME );
-
-				function checkResolving() {
-					const isResolving = window.wp.data
-						.select( 'core' )
-						.hasResolvingSelectors();
-
-					if ( isResolving ) {
-						clearTimeout( pauseTimeout );
-						pauseTimeout = undefined;
-						return;
-					}
-
-					if ( ! pauseTimeout ) {
-						pauseTimeout = setTimeout( finish, MAX_PAUSE_TIME );
-					}
-				}
-
-				const unsubscribe = window.wp.data.subscribe(
-					checkResolving,
-					'core'
-				);
-				checkResolving();
-			} );
-		} );
+		await this.waitForSiteEditor();
 	}
 }
